add header

+++ 0x03-python-data_structures/lists.h
@@ -18,7 +18,7 @@ typedef struct listint_s
 size_t print_listint(const listint_t *h);
 listint_t *add_nodeint_end(listint_t **head, const int n);
 void free_listint(listint_t *head);
-size_t len_listint(const listint_t *h);
+int len_listint(const listint_t *h);
 int is_palindrome(listint_t **head);
 
 #endif /* LISTS_H */
